ADC and DAC implementation
Was able to verify the resolution of the ADC by turning on each LED in turn as the resistor hit four divisions of the total 256 taps, 4 LEDS = 256/4 = 64.  NOTE 256 came from 2^8 based on 8-bit resolution. Was able to use the DAC to setup software trigger and output an analog waveform. My code is for sine, but I included all other waveforms and they can easily be called in.